github: revert to the default GitHub Actions shell to bash
As part of #10, I've set the default shell for the macos workflow to zsh.

Although modern macOS runs zsh as the default shell, we actually do
not suffer from reverting back to bash, but we do reduce the complexity,
which is always a benfit.

The runner's shell does not matter because all scripts in this repo are
executables that are executed by zsh due to the shabeng line.