github: continue the macos workflow when some dotfiles are missing
I'd like to print the dotfiles, but should some of them not exist, it
should not fail the entire workflow because it is not a true problem.

The inspection steps are just there to help maintainers, the missing files
will be created eventually. I'd rather merge the inspection steps early on,
than forget about them later on.

I am working on this intermitently, so I will most likely forget to inspect
those files otherwise.