macOS: initialise Zsh dotfiles with explanatory comments
Should those files not already exist, this installation script creates
them as a no-op. It takes the opportunity to open with an explanatory
comment about each of those files.

Installaing Homebrew modifies the zprofile file, so the empty dotfiles
must be craeted beforehand. To accomodate this order, the filenames were
changed accordingly.