macOS: remove default rebase.updateRefs configuration
When 'true' the setting causes the 'main' branch to update during
the command 'git pull origin main --rebase' (on feature branches).
This behaviour is undesired because the local 'main' should not be
rebased in that case.

I did use set setting to 'true' on Git for Windows, and the 'main'
branch did not update automatically during the same pull command.